lint & cleanup
