--- Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "deno_sdl2"
-version = "0.3.1"
-edition = "2018"
-description = "SDL2 Bindings to Deno"
-license = "MIT"
-repository = "https://github.com/littledivy/deno_sdl2"
-
-[features]
-default = []
-use-vcpkg = ["sdl2/use-vcpkg", "sdl2/static-link"]
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-sdl2 = { version = "0.34.5", features = ["ttf", "image", "unsafe_textures"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-deno_bindgen = "0.5.0"
-anyhow = "1.0.42"
-rodio = "0.14.0"
-lazy_static = "1.4.0"
-
-[package.metadata.vcpkg]
-dependencies = ["sdl2", "sdl2-image[libjpeg-turbo,tiff,libwebp]", "sdl2-ttf", "sdl2-gfx"]
-git = "https://github.com/microsoft/vcpkg"
-rev = "025e564979cc01d0fbc5c920aa8a36635efb01bb"
-
-[package.metadata.vcpkg.target]
-x86_64-pc-windows-msvc = { triplet = "x64-windows-static-md", dependencies = ["yasm", "sdl2", "sdl2-image[libjpeg-turbo,tiff,libwebp]", "sdl2-ttf", "sdl2-gfx"] }
